Don't use ChannelMultiplexer in RPCProtocol
Fixes #13960

Contributed on behalf of STMicroelectronics
